commit 2

+++ src/main/java/com/example/application/repository/ArtistRepository.java
@@ -12,8 +12,9 @@ import java.util.List;
 import java.util.Set;
 
 @Repository
-public interface ArtistRepository extends JpaRepository<Artist, Long>, JpaSpecificationExecutor<Artist> {
+public interface ArtistRepository extends JpaRepository<Artist, Integer>, JpaSpecificationExecutor<Artist> {
 
     @Query("select b from Artist b where b.name = :name or b.startActivityYear <= :year and b.endActivityYear >= :year or b.genres = :genres")
     List<Artist> filter(@RequestParam("name") String name,@RequestParam("year") int year,@RequestParam("genres") Integer[]genres);
+
 }
